requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
openai
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu
